adds more module CSS files and reorganizes project
