+++ src/component/AddTodoForm.module.css
@@ -1,0 +1,6 @@
+.addTodoForm {
+    display: flex;
+    gap: 10px;
+    align-items: baseline;
+    justify-content: center;
+}
