+++ src/component/TodoList.jsx
@@ -1,8 +1,9 @@
 import TodoListItem from './TodoListItem';
+import style from './TodoListItem.module.css';
 
 function TodoList({ todoList, onRemoveTodo }) {
 	return (
-		<ul className='todo-list'>
+		<ul className={style.todoList}>
 			{todoList.map(({ id, title }) => (
 				<TodoListItem
 					key={id}
