+++ src/component/TodoList.module.css
@@ -1,0 +1,5 @@
+.todoList {
+    list-style-type: none;
+    padding: 0;
+    margin: 0;
+}
