+++ src/component/TodoListItem.module.css
@@ -1,4 +1,8 @@
 .listItem {
+    display: flex;
+    gap: 10px;
+    align-items: baseline;
+    justify-content: center;
     margin-bottom: 5px;
     /* Solely for the sake of messing with interesting functions in CSS */
     color: color-mix(in hsl increasing hue, blanchedalmond 45%, coral 90%);
